Move live NASR discovery tests to the integration suite.
Keep unit NasrParseTest offline-only so CI unit runs are not blocked on NFDC
latency when discovery probes would eventually skip.
